refactor: classify tree dispatch, helper, query
